build-opencv is now independent of cwd
Turns out that build-opencv script only built properly when in bamboo
directory. With this change the script should be able to be called from
wherever and the script will work as intended.

The build-opencv script executes cmake in a build directory that is
created at the root of the git repository. We want to make sure that
this directory structure is always enforced.